Move to a separate file the matter of commit annotation.
