+++ release-controller/google_docs.py
@@ -11,7 +11,7 @@ from pydrive2.drive import GoogleDrive
 from const import OsKind, GUESTOS
 from git_repo import GitRepo
 from release_notes import PreparedReleaseNotes
-from commit_annotator import LocalCommitChangeDeterminator
+from commit_annotation import LocalCommitChangeDeterminator
 
 md = markdown.Markdown(
     extensions=["pymdownx.tilde", "pymdownx.details"],
